+++ src/mqtt.js
@@ -41,8 +41,9 @@ class MQTTCLIENT extends EventEmitter {
 
         try {
             const fullTopic = `${this.mqttPrefix}/${topic}`;
-            await this.mqttClient.publish(fullTopic, message);
-            const emitDebug = this.mqttDebug ? this.emit('debug', `MQTT publish: ${fullTopic}: ${message}`) : false;
+            const publishMessage = JSON.stringify(message, null, 2);
+            await this.mqttClient.publish(fullTopic, publishMessage);
+            const emitDebug = this.mqttDebug ? this.emit('debug', `MQTT publish: ${fullTopic}: ${publishMessage}`) : false;
         } catch (error) {
             this.emit('error', `MQTT Publish error: ${error}`);
         };